Refactor overload-characterization, move functionality to flex-module
